trying to propely split it all up
